feat/OPS-270 : 스페이스 로직에 사용자 정보 체크 추가

+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/space/controller/ApiV1SpaceController.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.tuna.zoopzoop.backend.domain.member.entity.Member;
 import org.tuna.zoopzoop.backend.domain.space.membership.entity.Membership;
+import org.tuna.zoopzoop.backend.domain.space.membership.enums.Authority;
 import org.tuna.zoopzoop.backend.domain.space.membership.enums.JoinState;
 import org.tuna.zoopzoop.backend.domain.space.membership.service.MembershipService;
 import org.tuna.zoopzoop.backend.domain.space.space.dto.ReqBodyForSpaceSave;
@@ -19,6 +20,7 @@ import org.tuna.zoopzoop.backend.domain.space.space.service.SpaceService;
 import org.tuna.zoopzoop.backend.global.rsData.RsData;
 import org.tuna.zoopzoop.backend.global.security.jwt.CustomUserDetails;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -34,9 +36,14 @@ public class ApiV1SpaceController {
     @PostMapping
     @Operation(summary = "스페이스 생성")
     public RsData<ResBodyForSpaceSave> createClub(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
             @Valid @RequestBody ReqBodyForSpaceSave reqBody
     ){
         Space newSpace = spaceService.createSpace(reqBody.name());
+
+        // ADMIN으로 입력
+        Member member = userDetails.getMember();
+        membershipService.addMemberToSpace(member, newSpace, Authority.ADMIN);
 
         return new RsData<>(
                 "201",
@@ -51,9 +58,16 @@ public class ApiV1SpaceController {
     @DeleteMapping("/{spaceId}")
     @Operation(summary = "스페이스 삭제")
     public RsData<Void> deleteSpace(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
             @PathVariable Integer spaceId
-    ){
+    ) throws AccessDeniedException {
+        // ADMIN 권한 체크
+        Member member = userDetails.getMember();
+        if(!membershipService.isMemberAdminInSpace(member, spaceService.getSpaceById(spaceId)))
+            throw new AccessDeniedException("스페이스의 ADMIN 권한이 필요합니다.");
+
         String deletedSpaceName = spaceService.deleteSpace(spaceId);
+
         return new RsData<>(
                 "200",
                 String.format("%s - 스페이스가 삭제됐습니다.", deletedSpaceName),
@@ -64,9 +78,15 @@ public class ApiV1SpaceController {
     @PutMapping("/{spaceId}")
     @Operation(summary = "스페이스 이름 변경")
     public RsData<ResBodyForSpaceSave> updateSpaceName(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
             @PathVariable Integer spaceId,
             @Valid @RequestBody ReqBodyForSpaceSave reqBody
-    ) {
+    ) throws AccessDeniedException {
+        // ADMIN 권한 체크
+        Member member = userDetails.getMember();
+        if(!membershipService.isMemberAdminInSpace(member, spaceService.getSpaceById(spaceId)))
+            throw new AccessDeniedException("스페이스의 ADMIN 권한이 필요합니다.");
+
         Space updatedSpace = spaceService.updateSpaceName(spaceId, reqBody.name());
 
         return new RsData<>(
